refactor: extract shared config from the tracer config
This change proposes refactoring the configuration object for the
tracer by moving tracer-specific sources under the new tracing
submodule. The existing config module is used for the shared
parameters, whereas the new ddtrace.tracing.config module houses
the tracer-specific configuration.
