+++ ddtrace/__init__.py
@@ -1,9 +1,9 @@
 from ._monkey import patch  # noqa: E402
 from ._monkey import patch_all
 from .pin import Pin  # noqa: E402
-from .settings import _config as config  # noqa: E402
 from .span import Span  # noqa: E402
-from .tracer import Tracer  # noqa: E402
+from .tracing.config import config  # noqa: E402
+from .tracing.tracer import Tracer  # noqa: E402
 from .version import get_version
