+++ ddtrace/tracing/config.py
@@ -1,0 +1,280 @@
+from copy import deepcopy
+import os
+from typing import List
+from typing import Optional
+from typing import Tuple
+
+from ddtrace.internal.utils.cache import cachedmethod
+from ddtrace.settings.config import config as shared_config
+from ddtrace.settings.http import HttpConfig
+from ddtrace.settings.integration import IntegrationConfig
+
+from ..internal.constants import PROPAGATION_STYLE_ALL
+from ..internal.constants import PROPAGATION_STYLE_DATADOG
+from ..internal.logger import get_logger
+from ..internal.utils.formats import asbool
+from ..internal.utils.formats import parse_tags_str
+from ..pin import Pin
+
+
+log = get_logger(__name__)
+
+
+def _parse_propagation_styles(name, default):
+    # type: (str, str) -> set[str]
+    """Helper to parse http propagation extract/inject styles via env variables.
+
+    The expected format is::
+
+        <style>[,<style>...]
+
+
+    The allowed values are:
+
+    - "datadog"
+    - "b3"
+    - "b3 single header"
+
+
+    The default value is ``"datadog"``.
+
+
+    Examples::
+
+        # Extract trace context from "x-datadog-*" or "x-b3-*" headers from upstream headers
+        DD_TRACE_PROPAGATION_STYLE_EXTRACT="datadog,b3"
+
+        # Inject the "b3: *" header into downstream requests headers
+        DD_TRACE_PROPAGATION_STYLE_INJECT="b3 single header"
+    """
+    styles = set()
+    envvar = os.getenv(name, default=default)
+    for style in envvar.split(","):
+        style = style.strip().lower()
+        if not style:
+            continue
+        if style not in PROPAGATION_STYLE_ALL:
+            raise ValueError(
+                "Unknown style {!r} provided for {!r}, allowed values are {!r}".format(
+                    style, name, PROPAGATION_STYLE_ALL
+                )
+            )
+        styles.add(style)
+    return styles
+
+
+# Borrowed from: https://stackoverflow.com/questions/20656135/python-deep-merge-dictionary-data#20666342
+def _deepmerge(source, destination):
+    """
+    Merge the first provided ``dict`` into the second.
+
+    :param dict source: The ``dict`` to merge into ``destination``
+    :param dict destination: The ``dict`` that should get updated
+    :rtype: dict
+    :returns: ``destination`` modified
+    """
+    for key, value in source.items():
+        if isinstance(value, dict):
+            # get node or create one
+            node = destination.setdefault(key, {})
+            _deepmerge(value, node)
+        else:
+            destination[key] = value
+
+    return destination
+
+
+def get_error_ranges(error_range_str):
+    # type: (str) -> List[Tuple[int, int]]
+    error_ranges = []
+    error_range_str = error_range_str.strip()
+    error_ranges_str = error_range_str.split(",")
+    for error_range in error_ranges_str:
+        values = error_range.split("-")
+        try:
+            # Note: mypy does not like variable type changing
+            values = [int(v) for v in values]  # type: ignore[misc]
+        except ValueError:
+            log.exception("Error status codes was not a number %s", values)
+            continue
+        error_range = (min(values), max(values))  # type: ignore[assignment]
+        error_ranges.append(error_range)
+    return error_ranges  # type: ignore[return-value]
+
+
+class TracerConfig(object):
+    """Tracer configuration object.
+
+    It exposes an API to set and retrieve global settings for each integration.
+    All integrations must use this instance to register their defaults, so that
+    they're public available and can be updated by users.
+    """
+
+    class _HTTPServerConfig(object):
+        _error_statuses = "500-599"  # type: str
+        _error_ranges = get_error_ranges(_error_statuses)  # type: List[Tuple[int, int]]
+
+        @property
+        def error_statuses(self):
+            # type: () -> str
+            return self._error_statuses
+
+        @error_statuses.setter
+        def error_statuses(self, value):
+            # type: (str) -> None
+            self._error_statuses = value
+            self._error_ranges = get_error_ranges(value)
+            # Mypy can't catch cached method's invalidate()
+            self.is_error_code.invalidate()  # type: ignore[attr-defined]
+
+        @property
+        def error_ranges(self):
+            # type: () -> List[Tuple[int, int]]
+            return self._error_ranges
+
+        @cachedmethod()
+        def is_error_code(self, status_code):
+            # type: (int) -> bool
+            """Returns a boolean representing whether or not a status code is an error code.
+            Error status codes by default are 500-599.
+            You may also enable custom error codes::
+
+                from ddtrace import config
+                config.http_server.error_statuses = '401-404,419'
+
+            Ranges and singular error codes are permitted and can be separated using commas.
+            """
+            for error_range in self.error_ranges:
+                if error_range[0] <= status_code <= error_range[1]:
+                    return True
+            return False
+
+    def __init__(self):
+        # use a dict as underlying storing mechanism
+        self._config = {}
+
+        header_tags = parse_tags_str(os.getenv("DD_TRACE_HEADER_TAGS", ""))
+        self.http = HttpConfig(header_tags=header_tags)
+
+        # Master switch for turning on and off trace search by default
+        # this weird invocation of getenv is meant to read the DD_ANALYTICS_ENABLED
+        # legacy environment variable. It should be removed in the future
+        legacy_config_value = os.getenv("DD_ANALYTICS_ENABLED", default=False)
+
+        self.analytics_enabled = asbool(os.getenv("DD_TRACE_ANALYTICS_ENABLED", default=legacy_config_value))
+
+        self.http_server = self._HTTPServerConfig()
+
+        self.logs_injection = asbool(os.getenv("DD_LOGS_INJECTION", default=False))
+
+        self.report_hostname = asbool(os.getenv("DD_TRACE_REPORT_HOSTNAME", default=False))
+
+        self.health_metrics_enabled = asbool(os.getenv("DD_TRACE_HEALTH_METRICS_ENABLED", default=False))
+
+        # Propagation styles
+        self._propagation_style_extract = _parse_propagation_styles(
+            "DD_TRACE_PROPAGATION_STYLE_EXTRACT", default=PROPAGATION_STYLE_DATADOG
+        )
+        self._propagation_style_inject = _parse_propagation_styles(
+            "DD_TRACE_PROPAGATION_STYLE_INJECT", default=PROPAGATION_STYLE_DATADOG
+        )
+
+        # Raise certain errors only if in testing raise mode to prevent crashing in production with non-critical errors
+        self._trace_compute_stats = asbool(os.getenv("DD_TRACE_COMPUTE_STATS", False))
+
+    def __getattr__(self, name):
+        try:
+            return self._config[name]
+        except KeyError:
+            try:
+                val = getattr(shared_config, name)
+            except AttributeError:
+                val = self._config[name] = IntegrationConfig(self, name)
+            return val
+
+    def get_from(self, obj):
+        """Retrieves the configuration for the given object.
+        Any object that has an attached `Pin` must have a configuration
+        and if a wrong object is given, an empty `dict` is returned
+        for safety reasons.
+        """
+        pin = Pin.get_from(obj)
+        if pin is None:
+            log.debug("No configuration found for %s", obj)
+            return {}
+
+        return pin._config
+
+    def _add(self, integration, settings, merge=True):
+        """Internal API that registers an integration with given default
+        settings.
+
+        :param str integration: The integration name (i.e. `requests`)
+        :param dict settings: A dictionary that contains integration settings;
+            to preserve immutability of these values, the dictionary is copied
+            since it contains integration defaults.
+        :param bool merge: Whether to merge any existing settings with those provided,
+            or if we should overwrite the settings with those provided;
+            Note: when merging existing settings take precedence.
+        """
+        # DEV: Use `getattr()` to call our `__getattr__` helper
+        existing = getattr(self, integration)
+        settings = deepcopy(settings)
+
+        if merge:
+            # DEV: This may appear backwards keeping `existing` as the "source" and `settings` as
+            #   the "destination", but we do not want to let `_add(..., merge=True)` overwrite any
+            #   existing settings
+            #
+            # >>> config.requests['split_by_domain'] = True
+            # >>> config._add('requests', dict(split_by_domain=False))
+            # >>> config.requests['split_by_domain']
+            # True
+            self._config[integration] = IntegrationConfig(self, integration, _deepmerge(existing, settings))
+        else:
+            self._config[integration] = IntegrationConfig(self, integration, settings)
+
+    def trace_headers(self, whitelist):
+        """
+        Registers a set of headers to be traced at global level or integration level.
+        :param whitelist: the case-insensitive list of traced headers
+        :type whitelist: list of str or str
+        :return: self
+        :rtype: HttpConfig
+        """
+        self.http.trace_headers(whitelist)
+        return self
+
+    def header_is_traced(self, header_name):
+        # type: (str) -> bool
+        """
+        Returns whether or not the current header should be traced.
+        :param header_name: the header name
+        :type header_name: str
+        :rtype: bool
+        """
+        return self.http.header_is_traced(header_name)
+
+    def _header_tag_name(self, header_name):
+        # type: (str) -> Optional[str]
+        return self.http._header_tag_name(header_name)
+
+    def _get_service(self, default=None):
+        """
+        Returns the globally configured service or the default if none is configured.
+
+        :param default: the default service to use if none is configured or
+            found.
+        :type default: str
+        :rtype: str|None
+        """
+        # TODO: This method can be replaced with `config.service`.
+        return self.service if self.service is not None else default
+
+    def __repr__(self):
+        cls = self.__class__
+        integrations = ", ".join(self._config.keys())
+        return "{}.{}({})".format(cls.__module__, cls.__name__, integrations)
+
+
+config = TracerConfig()
